+++ source/Dialogs/SaveProjectFileDialog.Designer.vb
@@ -40,6 +40,7 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_File = New System.Windows.Forms.GroupBox()
         Me.GroupBox_Options = New System.Windows.Forms.GroupBox()
         Me.SaveFileDialog1 = New System.Windows.Forms.SaveFileDialog()
+        Me.CheckBox_RelativePaths = New System.Windows.Forms.CheckBox()
         Me.GroupBox_DisplayOptions.SuspendLayout()
         Me.GroupBox_SeriesProperties.SuspendLayout()
         Me.GroupBox_File.SuspendLayout()
@@ -50,7 +51,7 @@ Partial Class SaveProjectFileDialog
         '
         Me.Button_OK.Anchor = CType((System.Windows.Forms.AnchorStyles.Bottom Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.Button_OK.DialogResult = System.Windows.Forms.DialogResult.OK
-        Me.Button_OK.Location = New System.Drawing.Point(316, 316)
+        Me.Button_OK.Location = New System.Drawing.Point(126, 246)
         Me.Button_OK.Name = "Button_OK"
         Me.Button_OK.Size = New System.Drawing.Size(75, 23)
         Me.Button_OK.TabIndex = 1
@@ -61,7 +62,7 @@ Partial Class SaveProjectFileDialog
         '
         Me.Button_Cancel.Anchor = CType((System.Windows.Forms.AnchorStyles.Bottom Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.Button_Cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel
-        Me.Button_Cancel.Location = New System.Drawing.Point(397, 316)
+        Me.Button_Cancel.Location = New System.Drawing.Point(207, 246)
         Me.Button_Cancel.Name = "Button_Cancel"
         Me.Button_Cancel.Size = New System.Drawing.Size(75, 23)
         Me.Button_Cancel.TabIndex = 5
@@ -74,9 +75,9 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_DisplayOptions.Controls.Add(Me.CheckBox_LineWidth)
         Me.GroupBox_DisplayOptions.Controls.Add(Me.CheckBox_LineStyle)
         Me.GroupBox_DisplayOptions.Controls.Add(Me.CheckBox_Color)
-        Me.GroupBox_DisplayOptions.Location = New System.Drawing.Point(9, 114)
+        Me.GroupBox_DisplayOptions.Location = New System.Drawing.Point(138, 42)
         Me.GroupBox_DisplayOptions.Name = "GroupBox_DisplayOptions"
-        Me.GroupBox_DisplayOptions.Size = New System.Drawing.Size(200, 116)
+        Me.GroupBox_DisplayOptions.Size = New System.Drawing.Size(123, 120)
         Me.GroupBox_DisplayOptions.TabIndex = 7
         Me.GroupBox_DisplayOptions.TabStop = False
         Me.GroupBox_DisplayOptions.Text = "Display options"
@@ -156,9 +157,9 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_SeriesProperties.Controls.Add(Me.CheckBox_Title)
         Me.GroupBox_SeriesProperties.Controls.Add(Me.CheckBox_Unit)
         Me.GroupBox_SeriesProperties.Controls.Add(Me.CheckBox_Interpretation)
-        Me.GroupBox_SeriesProperties.Location = New System.Drawing.Point(9, 19)
+        Me.GroupBox_SeriesProperties.Location = New System.Drawing.Point(9, 42)
         Me.GroupBox_SeriesProperties.Name = "GroupBox_SeriesProperties"
-        Me.GroupBox_SeriesProperties.Size = New System.Drawing.Size(200, 89)
+        Me.GroupBox_SeriesProperties.Size = New System.Drawing.Size(123, 120)
         Me.GroupBox_SeriesProperties.TabIndex = 8
         Me.GroupBox_SeriesProperties.TabStop = False
         Me.GroupBox_SeriesProperties.Text = "Series properties"
@@ -169,13 +170,13 @@ Partial Class SaveProjectFileDialog
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
         Me.TextBox_File.Location = New System.Drawing.Point(62, 15)
         Me.TextBox_File.Name = "TextBox_File"
-        Me.TextBox_File.Size = New System.Drawing.Size(352, 20)
+        Me.TextBox_File.Size = New System.Drawing.Size(162, 20)
         Me.TextBox_File.TabIndex = 9
         '
         'Button_Browse
         '
         Me.Button_Browse.Anchor = CType((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
-        Me.Button_Browse.Location = New System.Drawing.Point(420, 14)
+        Me.Button_Browse.Location = New System.Drawing.Point(230, 14)
         Me.Button_Browse.Name = "Button_Browse"
         Me.Button_Browse.Size = New System.Drawing.Size(34, 23)
         Me.Button_Browse.TabIndex = 10
@@ -200,7 +201,7 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_File.Controls.Add(Me.Button_Browse)
         Me.GroupBox_File.Location = New System.Drawing.Point(12, 12)
         Me.GroupBox_File.Name = "GroupBox_File"
-        Me.GroupBox_File.Size = New System.Drawing.Size(460, 46)
+        Me.GroupBox_File.Size = New System.Drawing.Size(270, 46)
         Me.GroupBox_File.TabIndex = 12
         Me.GroupBox_File.TabStop = False
         '
@@ -209,20 +210,31 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_Options.Anchor = CType((((System.Windows.Forms.AnchorStyles.Top Or System.Windows.Forms.AnchorStyles.Bottom) _
             Or System.Windows.Forms.AnchorStyles.Left) _
             Or System.Windows.Forms.AnchorStyles.Right), System.Windows.Forms.AnchorStyles)
+        Me.GroupBox_Options.Controls.Add(Me.CheckBox_RelativePaths)
         Me.GroupBox_Options.Controls.Add(Me.GroupBox_SeriesProperties)
         Me.GroupBox_Options.Controls.Add(Me.GroupBox_DisplayOptions)
         Me.GroupBox_Options.Location = New System.Drawing.Point(12, 64)
         Me.GroupBox_Options.Name = "GroupBox_Options"
-        Me.GroupBox_Options.Size = New System.Drawing.Size(460, 241)
+        Me.GroupBox_Options.Size = New System.Drawing.Size(270, 171)
         Me.GroupBox_Options.TabIndex = 13
         Me.GroupBox_Options.TabStop = False
         Me.GroupBox_Options.Text = "Options"
+        '
+        'CheckBox_RelativePaths
+        '
+        Me.CheckBox_RelativePaths.AutoSize = True
+        Me.CheckBox_RelativePaths.Location = New System.Drawing.Point(15, 19)
+        Me.CheckBox_RelativePaths.Name = "CheckBox_RelativePaths"
+        Me.CheckBox_RelativePaths.Size = New System.Drawing.Size(117, 17)
+        Me.CheckBox_RelativePaths.TabIndex = 9
+        Me.CheckBox_RelativePaths.Text = "Save relative paths"
+        Me.CheckBox_RelativePaths.UseVisualStyleBackColor = True
         '
         'SaveProjectFileDialog
         '
         Me.AutoScaleDimensions = New System.Drawing.SizeF(6.0!, 13.0!)
         Me.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font
-        Me.ClientSize = New System.Drawing.Size(484, 351)
+        Me.ClientSize = New System.Drawing.Size(294, 281)
         Me.Controls.Add(Me.GroupBox_Options)
         Me.Controls.Add(Me.GroupBox_File)
         Me.Controls.Add(Me.Button_OK)
@@ -230,7 +242,7 @@ Partial Class SaveProjectFileDialog
         Me.Icon = CType(resources.GetObject("$this.Icon"), System.Drawing.Icon)
         Me.MaximizeBox = False
         Me.MinimizeBox = False
-        Me.MinimumSize = New System.Drawing.Size(260, 390)
+        Me.MinimumSize = New System.Drawing.Size(310, 320)
         Me.Name = "SaveProjectFileDialog"
         Me.ShowInTaskbar = False
         Me.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent
@@ -242,6 +254,7 @@ Partial Class SaveProjectFileDialog
         Me.GroupBox_File.ResumeLayout(False)
         Me.GroupBox_File.PerformLayout()
         Me.GroupBox_Options.ResumeLayout(False)
+        Me.GroupBox_Options.PerformLayout()
         Me.ResumeLayout(False)
 
     End Sub
@@ -262,4 +275,5 @@ Partial Class SaveProjectFileDialog
     Friend WithEvents GroupBox_File As GroupBox
     Friend WithEvents GroupBox_Options As GroupBox
     Friend WithEvents SaveFileDialog1 As SaveFileDialog
+    Friend WithEvents CheckBox_RelativePaths As CheckBox
 End Class